cloud-deployment/facenet-service/main-simple.py
#!/usr/bin/env python3
"""
FaceNet Service for AutoSOS - Simplified Version
Cloud-deployed facial recognition service without OpenCV dependencies
"""

import os
import time
import base64
import json
import logging
from typing import Dict, Any, Optional
from contextlib import asynccontextmanager

import numpy as np
import tensorflow as tf
from fastapi import FastAPI, File, UploadFile, Form, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse
import redis
import structlog
from prometheus_client import Counter, Histogram, generate_latest, CONTENT_TYPE_LATEST
from fastapi.responses import Response

# Configure structured logging
structlog.configure(
    processors=[
        structlog.stdlib.filter_by_level,
        structlog.stdlib.add_logger_name,
        structlog.stdlib.add_log_level,
        structlog.stdlib.PositionalArgumentsFormatter(),
        structlog.processors.TimeStamper(fmt="iso"),
        structlog.processors.StackInfoRenderer(),
        structlog.processors.format_exc_info,
        structlog.processors.UnicodeDecoder(),
        structlog.processors.JSONRenderer()
    ],
    context_class=dict,
    logger_factory=structlog.stdlib.LoggerFactory(),
    wrapper_class=structlog.stdlib.BoundLogger,
    cache_logger_on_first_use=True,
)

logger = structlog.get_logger()

# Prometheus metrics
REQUEST_COUNT = Counter('facenet_requests_total', 'Total FaceNet requests', ['method', 'endpoint', 'status'])
REQUEST_DURATION = Histogram('facenet_request_duration_seconds', 'FaceNet request duration', ['method', 'endpoint'])

# Configuration
REDIS_URL = os.getenv("REDIS_URL", "redis://redis:6379")
MODEL_CACHE_DIR = os.getenv("MODEL_CACHE_DIR", "/app/models")
SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_KEY")

# Initialize Redis
redis_client = redis.from_url(REDIS_URL, decode_responses=True)

# Global FaceNet service instance
facenet_service = None

@asynccontextmanager
async def lifespan(app: FastAPI):
    """Application lifespan management"""
    global facenet_service
    
    logger.info("Starting FaceNet Service (Simplified)")
    
    try:
        # Initialize a simple FaceNet service
        facenet_service = SimpleFaceNetService()
        await facenet_service.initialize()
        logger.info("FaceNet service initialized successfully")
    except Exception as e:
        logger.error("Failed to initialize FaceNet service", error=str(e))
        # Don't raise, just log the error
    
    yield
    
    logger.info("Shutting down FaceNet Service")

class SimpleFaceNetService:
    """Simplified FaceNet service for testing"""
    
    def __init__(self):
        self.model = None
        self.face_database = {}
        
    async def initialize(self):
        """Initialize the service"""
        logger.info("Initializing simplified FaceNet service")
        # Create a simple model for testing
        self.model = tf.keras.Sequential([
            tf.keras.layers.Dense(128, activation='relu', input_shape=(224*224*3,)),
            tf.keras.layers.Dense(64, activation='relu'),
            tf.keras.layers.Dense(32, activation='linear')
        ])
        logger.info("Simplified FaceNet model created")
        
    def get_face_count(self):
        """Get the number of registered faces"""
        return len(self.face_database)

# Initialize FastAPI app
app = FastAPI(
    title="AutoSOS FaceNet Service (Simplified)",
    description="Simplified facial recognition service for AutoSOS",
    version="1.0.0",
    lifespan=lifespan
)

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Configure properly for production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

@app.get("/health")
async def health_check():
    """Service health check"""
    if facenet_service is None:
        return JSONResponse(
            status_code=503,
            content={"status": "unhealthy", "reason": "FaceNet service not initialized"}
        )
    
    return {
        "status": "healthy",
        "service": "facenet",
        "timestamp": time.time(),
        "face_count": facenet_service.get_face_count(),
        "version": "simplified"
    }

@app.get("/metrics")
async def metrics():
    """Prometheus metrics endpoint"""
    return Response(generate_latest(), media_type=CONTENT_TYPE_LATEST)

@app.post("/register-face")
async def register_face(
    user_id: str = Form(...),
    user_name: str = Form(...),
    file: UploadFile = File(...)
):
    """Register a new face for recognition"""
    REQUEST_COUNT.labels(method="POST", endpoint="/register-face", status="200").inc()
    
    if facenet_service is None:
        raise HTTPException(status_code=503, detail="FaceNet service not initialized")
    
    try:
        # Read image data
        image_data = await file.read()
        
        # Store in database (simplified)
        facenet_service.face_database[user_id] = {
            "user_name": user_name,
            "registered_at": time.time(),
            "image_size": len(image_data)
        }
        
        logger.info("Face registered successfully", user_id=user_id, user_name=user_name)
        
        return {
            "success": True,
            "user_id": user_id,
            "user_name": user_name,
            "message": "Face registered successfully (simplified mode)"
        }
            
    except Exception as e:
        logger.error("Face registration failed", error=str(e), user_id=user_id)
        raise HTTPException(status_code=500, detail="Internal server error")

@app.post("/recognize-face")
async def recognize_face(file: UploadFile = File(...)):
    """Recognize a face from uploaded image"""
    REQUEST_COUNT.labels(method="POST", endpoint="/recognize-face", status="200").inc()
    
    if facenet_service is None:
        raise HTTPException(status_code=503, detail="FaceNet service not initialized")
    
    try:
        # Read image data
        image_data = await file.read()
        
        # Simple recognition logic (for testing)
        if len(facenet_service.face_database) > 0:
            # Return the first registered user (simplified)
            user_id = list(facenet_service.face_database.keys())[0]
            user_info = facenet_service.face_database[user_id]
            
            return {
                "success": True,
                "recognized": True,
                "user_id": user_id,
                "user_name": user_info["user_name"],
                "confidence": 0.95,
                "similarity": 0.95,
                "message": "Face recognized (simplified mode)"
            }
        else:
            return {
                "success": True,
                "recognized": False,
                "message": "No faces registered in database"
            }
            
    except Exception as e:
        logger.error("Face recognition failed", error=str(e))
        raise HTTPException(status_code=500, detail="Internal server error")

@app.get("/face-count")
async def get_face_count():
    """Get the number of registered faces"""
    if facenet_service is None:
        raise HTTPException(status_code=503, detail="FaceNet service not initialized")
    
    count = facenet_service.get_face_count()
    return {"face_count": count}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8001)
