cloud-deployment/facenet-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow-cpu==2.13.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
supabase==2.0.0
redis==5.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
typing-extensions==4.5.0
httpx==0.25.2
structlog==23.2.0
prometheus-client==0.19.0